Add workaround for julia 1.10 and CUDA
As of CUDA 5.1.1, there are some issues with ClimaAtmos and Julia 1.10.
It seems that the issue is with this particular function and the error
is due to added calles to gc_frame functions that cannot be compiled on
a GPU.

For unknown reasons, the code compiles when we remove the keyword
arguments. While the issue is being investigated, we can work around it
by excluding the offending branch